Minor changes and crying
Changed the computation of the deterministic x_true and merge it with the best lambda to be consistent. Changed the whole notebook to be in accordance. Very good result for the first plots but want to kill my self after the middle of the notebook. Changed the comparison to argmax to 2-norm because it was dumb b4 (my bad) but it changed nothing). I wanna cry...